Remove unused code and general file cleanup build_Creator3.0.149

+++ CreatorCode/CreatorData/ProductLicenseTable.cs
@@ -121,7 +121,7 @@ namespace CreatorData
             }
         }
 
-        public static int GetProductVersionFromTable(int productID)
+        public static int GetProductVersionFromTable(byte productID)
         {
             using (CreatorDataContext db = new CreatorDataContext())
             {
